Use compileSdk & targetSdk 35 (Android 15).

+++ app/src/main/java/nerd/tuxmobil/fahrplan/congress/system/OnBootReceiver.kt
@@ -33,20 +33,35 @@ class OnBootReceiver : BroadcastReceiver() {
                 logging = logging,
             )
         }
-        val alarmServices = AlarmServices.newInstance(context, appRepository)
+        configureSessionAlarms(context, appRepository, nowMoment)
+        configureScheduleUpdates(context, appRepository)
+    }
+
+    private fun configureSessionAlarms(
+        context: Context,
+        appRepository: AppRepository,
+        nowMoment: Moment,
+    ) {
         val alarms = appRepository.readAlarms()
-        for (alarm in alarms) {
-            // Check if the alarm time has passed
-            if (nowMoment.isBefore(alarm.startTime)) {
-                logging.d(LOG_TAG, "Scheduling alarm for session: ${alarm.sessionTitle}, ${alarm.sessionTitle}")
-                alarmServices.scheduleSessionAlarm(alarm.toSchedulableAlarm())
-            } else {
-                logging.d(LOG_TAG, "Deleting alarm from database: $alarm")
-                appRepository.deleteAlarmForAlarmId(alarm.id)
+        if (alarms.isNotEmpty()) {
+            val alarmServices = AlarmServices.newInstance(context, appRepository)
+            for (alarm in alarms) {
+                // Check if the alarm time has passed
+                if (nowMoment.isBefore(alarm.startTime)) {
+                    logging.d(LOG_TAG, "Scheduling alarm for session: ${alarm.sessionTitle}, ${alarm.sessionTitle}")
+                    alarmServices.scheduleSessionAlarm(alarm.toSchedulableAlarm())
+                } else {
+                    logging.d(LOG_TAG, "Deleting alarm from database: $alarm")
+                    appRepository.deleteAlarmForAlarmId(alarm.id)
+                }
             }
         }
+    }
 
-        // Start auto updates
+    private fun configureScheduleUpdates(
+        context: Context,
+        appRepository: AppRepository,
+    ) {
         if (appRepository.readAutoUpdateEnabled()) { // Check if auto update is enabled
             val lastFetchedAt = appRepository.readScheduleLastFetchedAt()
             val nowMillis = Moment.now().toMilliseconds()
